+++ activesupport/lib/active_support/callbacks.rb
@@ -6,7 +6,6 @@ require 'active_support/core_ext/kernel/reporting'
 require 'active_support/core_ext/kernel/singleton_class'
 require 'active_support/core_ext/module/attribute_accessors'
 require 'active_support/core_ext/string/filters'
-require 'active_support/deprecation'
 require 'thread'
 
 module ActiveSupport
@@ -66,12 +65,6 @@ module ActiveSupport
     end
 
     CALLBACK_FILTER_TYPES = [:before, :after, :around]
-
-    # If true, Active Record and Active Model callbacks returning +false+ will
-    # halt the entire callback chain and display a deprecation message.
-    # If false, callback chains will only be halted by calling +throw :abort+.
-    # Defaults to +true+.
-    mattr_accessor(:halt_and_display_warning_on_return_false, instance_writer: false) { true }
 
     # Runs the callbacks for the given event.
     #
@@ -295,13 +288,6 @@ module ActiveSupport
 
     class Callback #:nodoc:#
       def self.build(chain, filter, kind, options)
-        if filter.is_a?(String)
-          ActiveSupport::Deprecation.warn(<<-MSG.squish)
-            Passing string to define callback is deprecated and will be removed
-            in Rails 5.1 without replacement.
-          MSG
-        end
-
         new chain.name, filter, kind, options, chain.config
       end
 
@@ -767,25 +753,14 @@ module ActiveSupport
         Proc.new do |target, result_lambda|
           terminate = true
           catch(:abort) do
-            result = result_lambda.call if result_lambda.is_a?(Proc)
-            if Callbacks.halt_and_display_warning_on_return_false && result == false
-              display_deprecation_warning_for_false_terminator
-            else
-              terminate = false
-            end
+            result_lambda.call if result_lambda.is_a?(Proc)
+            terminate = false
           end
           terminate
         end
       end
 
       private
-
-      def display_deprecation_warning_for_false_terminator
-        ActiveSupport::Deprecation.warn(<<-MSG.squish)
-          Returning `false` in Active Record and Active Model callbacks will not implicitly halt a callback chain in Rails 5.1.
-          To explicitly halt the callback chain, please use `throw :abort` instead.
-        MSG
-      end
     end
   end
 end